The Big Configuration Rework (and some other things too)
