+++ arch/pico/impl/psx.c
@@ -6,6 +6,6 @@ PSX_Controller_t _psx = {
   .padding_5a = 0x5A
 };
 
-#if defined(PSX_DRIVER)
+#if defined(PSX_DRIVER) && defined(PSX_ACTIVE)
 #include INCLUDE_IMPL(psx, PSX_DRIVER, c)
 #endif
