+++ boards/community/mickey/protoboard/rules.mk
@@ -1,0 +1,1 @@
+TARGET_ARCH = pico
