--- configs/examples/arduino/pro-micro/iidx/default/config.h
@@ -1,10 +1,0 @@
-#pragma once
-// The Arduino Pro Micro is a general development board.
-// The following provides a sample user configuration for the Pro Micro dev board in a IIDX controller.
-
-// Our example controller requires no additional configuration, so our "default" config will be fairly empty.
-// However, other controller configs are available to e.g. enable PS2 support or use a different encoder PPR.
-
-/*** Additional Includes ***/
-// We'll include some friendly USB labels in a separate file.
-#include "labels.h"
