--- configs/examples/arduino/pro-micro/iidx/default/labels.h
@@ -1,35 +1,0 @@
-#pragma once
-// The following provides customizable USB strings to USBemani.
-// This allows for certain software to describe the functions in a more friendly manner.
-// Not every piece of software supports every type of string.
-
-/*** Buttons ***/
-#define USB_STRING_BUTTON_1   "Key 1"
-#define USB_STRING_BUTTON_2   "Key 2"
-#define USB_STRING_BUTTON_3   "Key 3"
-#define USB_STRING_BUTTON_4   "Key 4"
-#define USB_STRING_BUTTON_5   "Key 5"
-#define USB_STRING_BUTTON_6   "Key 6"
-#define USB_STRING_BUTTON_7   "Key 7"
-#define USB_STRING_BUTTON_8   "E1"
-#define USB_STRING_BUTTON_9   "E2"
-#define USB_STRING_BUTTON_10  "E3"
-#define USB_STRING_BUTTON_11  "E4"
-
-/*** Encoders ***/
-#define USB_STRING_ENCODER_1  "Turntable"
-
-/*** RGB ***/
-// Our example controller has 12 RGB channels: 7 keys, the 4 E- buttons, and the turntable (in that order)
-#define USB_STRING_RGB_1      "Key 1"
-#define USB_STRING_RGB_2      "Key 2"
-#define USB_STRING_RGB_3      "Key 3"
-#define USB_STRING_RGB_4      "Key 4"
-#define USB_STRING_RGB_5      "Key 5"
-#define USB_STRING_RGB_6      "Key 6"
-#define USB_STRING_RGB_7      "Key 7"
-#define USB_STRING_RGB_8      "E1"
-#define USB_STRING_RGB_9      "E2"
-#define USB_STRING_RGB_10     "E3"
-#define USB_STRING_RGB_11     "E4"
-#define USB_STRING_RGB_12     "Turntable Ring"
