--- configs/examples/arduino/pro-micro/iidx/default/rules.mk
@@ -1,1 +1,0 @@
-# This file is empty, but must be present
