--- configs/examples/arduino/pro-micro/iidx/with-overrides/user.c
@@ -1,5 +1,0 @@
-// The Arduino Pro Micro is a general development board.
-// The following provides a sample of user-customized code to USBemani.
-
-// We'll include the same code we're using for the "default" configuration.
-#include "../default/user.c"
