--- configs/examples/raspberrypi/pico/iidx/default/config.h
@@ -1,7 +1,0 @@
-#pragma once
-// The Raspberry Pi Pico is a general development board.
-// The following provides a sample user configuration for the Pico dev board in a IIDX controller.
-
-/*** Additional Includes ***/
-// We'll include some friendly USB labels in a separate file.
-#include "labels.h"
