--- configs/speedylabs/pocket-voltex/rev5/default/config.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "labels.h"
