--- configs/speedylabs/pocket-voltex/rev5/default/labels.h
@@ -1,10 +1,0 @@
-#pragma once
-#define USB_STRING_LIGHT_1 "BT-A"
-#define USB_STRING_LIGHT_2 "BT-B"
-#define USB_STRING_LIGHT_3 "BT-C"
-#define USB_STRING_LIGHT_4 "BT-D"
-#define USB_STRING_LIGHT_5 "FX-L"
-#define USB_STRING_LIGHT_6 "FX-R"
-#define USB_STRING_LIGHT_7 "Start"
-
-#define USB_STRING_RGB_1  "Controller"
