+++ layouts/community/mickey/djdao-fps/rules.mk
@@ -1,0 +1,2 @@
+TARGET_BOARD      ?= community/mickey/protoboard
+TARGET_CONTROLLER ?= djdao/iidx2012
