+++ layouts/default/speedylabs/pocket-voltex/rules.mk
@@ -1,0 +1,2 @@
+TARGET_BOARD      ?= speedylabs/pocket-voltex/rev5
+TARGET_CONTROLLER ?= speedylabs/pocket-voltex/rev5
