+++ layouts/examples/iidx/arduino/pro-micro/rules.mk
@@ -1,0 +1,2 @@
+TARGET_BOARD      ?= dev/arduino/pro-micro
+TARGET_CONTROLLER ?= examples/iidx/entry-model
