+++ layouts/examples/iidx/usb_labels.h
@@ -1,0 +1,32 @@
+#pragma once
+// The following provides a number of USB labels used by the examples in this folder.
+// Supporting software will use these labels instead of generic labels.
+
+/*** Buttons ***/
+#define USB_STRING_BUTTON_1   "Key 1"
+#define USB_STRING_BUTTON_2   "Key 2"
+#define USB_STRING_BUTTON_3   "Key 3"
+#define USB_STRING_BUTTON_4   "Key 4"
+#define USB_STRING_BUTTON_5   "Key 5"
+#define USB_STRING_BUTTON_6   "Key 6"
+#define USB_STRING_BUTTON_7   "Key 7"
+#define USB_STRING_BUTTON_8   "E1/Start"
+#define USB_STRING_BUTTON_9   "E2/Select"
+#define USB_STRING_BUTTON_10  "E3"
+#define USB_STRING_BUTTON_11  "E4"
+
+/*** Encoders ***/
+#define USB_STRING_ENCODER_1  "Turntable"
+
+/*** Lights ***/
+#define USB_STRING_LIGHT_1      "Key 1"
+#define USB_STRING_LIGHT_2      "Key 2"
+#define USB_STRING_LIGHT_3      "Key 3"
+#define USB_STRING_LIGHT_4      "Key 4"
+#define USB_STRING_LIGHT_5      "Key 5"
+#define USB_STRING_LIGHT_6      "Key 6"
+#define USB_STRING_LIGHT_7      "Key 7"
+#define USB_STRING_LIGHT_8      "E1/Start"
+#define USB_STRING_LIGHT_9      "E2/Select"
+#define USB_STRING_LIGHT_10     "E3"
+#define USB_STRING_LIGHT_11     "E4"
